Colocate related stories
