+++ src/components/NavBar/NavBar.story.jsx
@@ -2,7 +2,6 @@ import React from 'react'
 
 import { NavigationBar } from './NavBar'
 export default {
-	component: NavigationBar,
 	title: 'Components/Navigation',
 }
